syncsigjmp and syncsigwait
syncsigjmp demonstrates using `sigsetmp` and `siglongjmp` to jump
out of a signal handler and into a dedicated signal-handling thread,
where the signal mask is set such that we can safely invoke
`pthread_cond_broadcast` on a condition variable.

Note that we vector all signal delivery to this thread, so that we
know that we never interrupt a non-async-signal-safe function.

syncsigwait demonstrates vector all signals to a dedicated thread
that sits in a tight loop running `sigwait`, thus turning
asynchronous signal delivery a synchronous return from a blocking
call.  This allows us to easily translate signal delivery to
events on condition variables.

Signed-off-by: Dan Cross <cross@gajendra.net>